hotfix execute service and validate tasks

+++ src/services/execute.service.ts
@@ -1,33 +1,12 @@
-import { questionExecuteTask } from '../utils';
+import { questionExecuteDangerousTask, questionExecuteTask } from '../utils';
 import logger from '../libs/logger';
 import chalk from 'chalk';
-import { ITask } from '../types';
+import { CommandTask, ITask, TaskType, WriteFileTask } from '../types';
 import * as fs from 'fs';
 import { executeCommand } from '../utils/command';
 
 abstract class BaseExecuteService {
   private nextHandler: BaseExecuteService | null = null;
-  async execute(task: ITask, isAutoExecute: boolean): Promise<void> {
-    if (!isAutoExecute && !task.dangerous) {
-      const confirmation = await questionExecuteTask(task.description);
-      if (!confirmation) {
-        logger.info(`Skipping: ${chalk.green(task.description)}`);
-        return;
-      }
-    }
-
-    if (task.dangerous) {
-      const confirmation = await questionExecuteTask(
-        `⚠️ Execute dangerous task: ${chalk.red(task.description)}`,
-      );
-      if (!confirmation) {
-        logger.info(
-          `Skipping dangerous task: ${chalk.green(task.description)}`,
-        );
-        return;
-      }
-    }
-  }
   setNext(handler: BaseExecuteService): BaseExecuteService {
     this.nextHandler = handler;
     return handler;
@@ -41,10 +20,21 @@ abstract class BaseExecuteService {
 }
 
 class ExecuteCommandsService extends BaseExecuteService {
-  async execute(task: ITask, isAutoExecute: boolean): Promise<void> {
+  async execute(task: CommandTask, isAutoExecute: boolean): Promise<void> {
+    if (!isAutoExecute) {
+      const question = task.dangerous
+        ? questionExecuteDangerousTask
+        : questionExecuteTask;
+      const confirmation = await question(task.description);
+      if (!confirmation) {
+        logger.warn(`Skipping: ${chalk.green(task.description)}`);
+        return;
+      }
+    }
+
     const { error } = await executeCommand(task.command!);
     if (error) {
-      logger.error(`❌ ${task.description}`);
+      logger.error(`❌ ${task.description}: ${error}`);
       return;
     }
 
@@ -52,9 +42,8 @@ class ExecuteCommandsService extends BaseExecuteService {
   }
 
   async handle(task: ITask, isAutoExecute: boolean): Promise<void> {
-    await super.execute(task, isAutoExecute);
-    if (task.type === 'COMMAND') {
-      await this.execute(task, isAutoExecute);
+    if (task.type === TaskType.COMMAND) {
+      await this.execute(task as CommandTask, isAutoExecute);
       return;
     }
     return super.handle(task, isAutoExecute);
@@ -62,14 +51,24 @@ class ExecuteCommandsService extends BaseExecuteService {
 }
 
 class ExecuteWriteFileService extends BaseExecuteService {
-  async execute(task: ITask, isAutoExecute: boolean): Promise<void> {
+  async execute(task: WriteFileTask, isAutoExecute: boolean): Promise<void> {
+    if (!isAutoExecute) {
+      const question = task.dangerous
+        ? questionExecuteDangerousTask
+        : questionExecuteTask;
+      const confirmation = await question(task.description);
+      if (!confirmation) {
+        logger.info(`Skipping: ${chalk.green(task.description)}`);
+        return;
+      }
+    }
+
     fs.writeFileSync(task.path!, task.content!);
     logger.info(`✅ ${task.description}`);
   }
   async handle(task: ITask, isAutoExecute: boolean): Promise<void> {
-    await super.execute(task, isAutoExecute);
-    if (task.type === 'WRITE_FILE') {
-      await this.execute(task, isAutoExecute);
+    if (task.type === TaskType.WRITE_FILE) {
+      await this.execute(task as WriteFileTask, isAutoExecute);
       return;
     }
     return super.handle(task, isAutoExecute);
@@ -77,7 +76,6 @@ class ExecuteWriteFileService extends BaseExecuteService {
 }
 
 const ex = new ExecuteCommandsService();
-ex.setNext(new ExecuteCommandsService());
 ex.setNext(new ExecuteWriteFileService());
 
 export { ex };
